+++ ionoscloud/provider.go
@@ -67,7 +67,7 @@ func Provider() *schema.Provider {
 			FirewallResource:           resourceFirewall(),
 			LanResource:                resourceLan(),
 			"ionoscloud_loadbalancer":  resourceLoadbalancer(),
-			"ionoscloud_nic":           resourceNic(),
+			nicResource:                resourceNic(),
 			ServerResource:             resourceServer(),
 			VolumeResource:             resourceVolume(),
 			GroupResource:              resourceGroup(),
@@ -99,6 +99,7 @@ func Provider() *schema.Provider {
 			UserResource:               dataSourceUser(),
 			IpBLockResource:            dataSourceIpBlock(),
 			VolumeResource:             dataSourceVolume(),
+			nicResource:                dataSourceNIC(),
 		},
 	}
 	provider.ConfigureContextFunc = func(ctx context.Context, d *schema.ResourceData) (interface{}, diag.Diagnostics) {