+++ ionoscloud/resource_nic_test.go
@@ -10,7 +10,7 @@ import (
 	"github.com/hashicorp/terraform-plugin-sdk/v2/terraform"
 )
 
-func TestAccNic_Basic(t *testing.T) {
+func TestAccNicBasic(t *testing.T) {
 	var nic ionoscloud.Nic
 	volumeName := "volume"
 
@@ -22,43 +22,23 @@ func TestAccNic_Basic(t *testing.T) {
 		CheckDestroy:      testAccCheckNicDestroyCheck,
 		Steps: []resource.TestStep{
 			{
-				Config: fmt.Sprintf(testaccchecknicconfigBasic, volumeName),
+				Config: fmt.Sprintf(testAccCheckNicConfigBasic, volumeName),
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNICExists("ionoscloud_nic.database_nic", &nic),
-					testAccCheckNicAttributes("ionoscloud_nic.database_nic", volumeName),
-					resource.TestCheckResourceAttrSet("ionoscloud_nic.database_nic", "mac"),
-					resource.TestCheckResourceAttr("ionoscloud_nic.database_nic", "name", volumeName),
-					resource.TestCheckResourceAttr("ionoscloud_nic.database_nic", "dhcp", "false"),
+					testAccCheckNICExists(fullNicResourceName, &nic),
+					resource.TestCheckResourceAttr(fullNicResourceName, "name", volumeName),
+					resource.TestCheckResourceAttr(fullNicResourceName, "dhcp", "true"),
+					resource.TestCheckResourceAttrSet(fullNicResourceName, "mac"),
+					resource.TestCheckResourceAttr(fullNicResourceName, "firewall_active", "true"),
+					resource.TestCheckResourceAttrPair(fullNicResourceName, "ips.0", "ionoscloud_ipblock.test_server", "ips.0"),
+					resource.TestCheckResourceAttrPair(fullNicResourceName, "ips.1", "ionoscloud_ipblock.test_server", "ips.1"),
 				),
 			},
 			{
-				Config: testaccchecknicconfigUpdate,
+				Config: testAccCheckNicConfigUpdate,
 				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNicAttributes("ionoscloud_nic.database_nic", "updated"),
-					resource.TestCheckResourceAttr("ionoscloud_nic.database_nic", "name", "updated"),
-					resource.TestCheckResourceAttr("ionoscloud_nic.database_nic", "dhcp", "true"),
-				),
-			},
-		},
-	})
-}
-
-func TestAccNic_Ips(t *testing.T) {
-	var nic ionoscloud.Nic
-	resource.Test(t, resource.TestCase{
-		PreCheck: func() {
-			testAccPreCheck(t)
-		},
-		ProviderFactories: testAccProviderFactories,
-		CheckDestroy:      testAccCheckNicDestroyCheck,
-		Steps: []resource.TestStep{
-			{
-				Config: fmt.Sprintf(testaccchecknicconfigIps),
-				Check: resource.ComposeTestCheckFunc(
-					testAccCheckNICExists("ionoscloud_nic.database_nic", &nic),
-					resource.TestCheckResourceAttrSet("ionoscloud_nic.database_nic", "mac"),
-					resource.TestCheckResourceAttrPair("ionoscloud_nic.database_nic", "ips.0", "ionoscloud_ipblock.webserver", "ips.0"),
-					resource.TestCheckResourceAttrPair("ionoscloud_nic.database_nic", "ips.1", "ionoscloud_ipblock.webserver", "ips.1"),
+					resource.TestCheckResourceAttr(fullNicResourceName, "name", "updated"),
+					resource.TestCheckResourceAttr(fullNicResourceName, "dhcp", "false"),
+					resource.TestCheckResourceAttr(fullNicResourceName, "firewall_active", "false"),
 				),
 			},
 		},
@@ -74,7 +54,7 @@ func testAccCheckNicDestroyCheck(s *terraform.State) error {
 	}
 
 	for _, rs := range s.RootModule().Resources {
-		if rs.Type != "ionoscloud_nic" {
+		if rs.Type != nicResource {
 			continue
 		}
 
@@ -94,20 +74,6 @@ func testAccCheckNicDestroyCheck(s *terraform.State) error {
 	}
 
 	return nil
-}
-
-func testAccCheckNicAttributes(n string, name string) resource.TestCheckFunc {
-	return func(s *terraform.State) error {
-		rs, ok := s.RootModule().Resources[n]
-		if !ok {
-			return fmt.Errorf("testAccCheckNicAttributes: Not found: %s", n)
-		}
-		if rs.Primary.Attributes["name"] != name {
-			return fmt.Errorf("bad name: %s", rs.Primary.Attributes["name"])
-		}
-
-		return nil
-	}
 }
 
 func testAccCheckNICExists(n string, nic *ionoscloud.Nic) resource.TestCheckFunc {
@@ -147,52 +113,19 @@ func testAccCheckNICExists(n string, nic *ionoscloud.Nic) resource.TestCheckFunc
 	}
 }
 
-const testaccchecknicconfigBasic = `
-resource "ionoscloud_datacenter" "foobar" {
+const testCreateDataCenterAndServer = `
+resource "ionoscloud_datacenter" "test_datacenter" {
 	name       = "nic-test"
 	location = "us/las"
 }
-
-resource "ionoscloud_server" "webserver" {
-  name = "webserver"
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
-  cores = 1
-  ram = 1024
-  availability_zone = "ZONE_1"
-  cpu_family = "AMD_OPTERON"
-	image_name ="ubuntu-16.04"
-	image_password = "K3tTj8G14a3EgKyNeeiY"
-  volume {
-    name = "system"
-    size = 5
-    disk_type = "SSD"
-
+resource "ionoscloud_ipblock" "test_server" {
+  location = ionoscloud_datacenter.test_datacenter.location
+  size = 2
+  name = "test_server_ipblock"
 }
-  nic {
-    lan = "1"
-    dhcp = true
-    firewall_active = true
-  }
-}
-
-resource "ionoscloud_nic" "database_nic" {
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
-  server_id = "${ionoscloud_server.webserver.id}"
-  lan = 2
-  dhcp = false
-  firewall_active = true
-  name = "%s"
-}`
-
-const testaccchecknicconfigUpdate = `
-resource "ionoscloud_datacenter" "foobar" {
-	name       = "nic-test"
-	location = "us/las"
-}
-
-resource "ionoscloud_server" "webserver" {
-  name = "webserver"
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
+resource "ionoscloud_server" "test_server" {
+  name = "test_server"
+  datacenter_id = "${ionoscloud_datacenter.test_datacenter.id}"
   cores = 1
   ram = 1024
   availability_zone = "ZONE_1"
@@ -209,58 +142,29 @@ resource "ionoscloud_server" "webserver" {
     dhcp = true
     firewall_active = true
   }
-}
-
-resource "ionoscloud_nic" "database_nic" {
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
-  server_id = "${ionoscloud_server.webserver.id}"
-  lan = 2
-  dhcp = true
-  firewall_active = true
-  name = "updated"
 }
 `
 
-const testaccchecknicconfigIps = `
-resource "ionoscloud_datacenter" "foobar" {
-	name       = "nic-test"
-	location = "us/las"
-}
-
-resource "ionoscloud_ipblock" "webserver" {
-  location = ionoscloud_datacenter.foobar.location
-  size = 2
-  name = "webserver_ipblock"
-}
-
-resource "ionoscloud_server" "webserver" {
-  name = "webserver"
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
-  cores = 1
-  ram = 1024
-  availability_zone = "ZONE_1"
-  cpu_family = "AMD_OPTERON"
-	image_name ="ubuntu-16.04"
-	image_password = "K3tTj8G14a3EgKyNeeiY"
-  volume {
-    name = "system"
-    size = 5
-    disk_type = "SSD"
-
-}
-  nic {
-    lan = "1"
-    dhcp = true
-    firewall_active = true
-  }
-}
-
+const testAccCheckNicConfigBasic = testCreateDataCenterAndServer + `
 resource "ionoscloud_nic" "database_nic" {
-  datacenter_id = "${ionoscloud_datacenter.foobar.id}"
-  server_id = "${ionoscloud_server.webserver.id}"
+  datacenter_id = ` + DatacenterResource + `.` + DatacenterTestResource + `.id
+  server_id = ` + ServerResource + `.` + ServerTestResource + `.id
+  lan = 2
+  firewall_active = true
+  dhcp = true
+  ips            = [ ionoscloud_ipblock.test_server.ips[0], ionoscloud_ipblock.test_server.ips[1] ]
+  name = "%s"
+}
+`
+
+const testAccCheckNicConfigUpdate = testCreateDataCenterAndServer + `
+resource "ionoscloud_nic" "database_nic" {
+  datacenter_id = ` + DatacenterResource + `.` + DatacenterTestResource + `.id
+  server_id = ` + ServerResource + `.` + ServerTestResource + `.id
   lan = 2
   dhcp = false
-  firewall_active = true
-  ips            = [ ionoscloud_ipblock.webserver.ips[0], ionoscloud_ipblock.webserver.ips[1] ]
-  name = "test_nic"
-}`
+  firewall_active = false
+  ips            = [ ionoscloud_ipblock.test_server.ips[0], ionoscloud_ipblock.test_server.ips[1] ]
+  name = "updated"
+}
+`